compatible(integration_test_charm.yaml): Add juju version prefix
In nested CI calls (e.g. release.yaml calls ci.yaml calls this workflow) the job name in ci.yaml will not show up on the GitHub Actions sidebar.

If this workflow is called with a matrix (e.g. to test multiple juju versions), the ci.yaml job name containing the Juju version will be lost.

So, we add the Juju version to one of the first jobs in this workflow.

(In the UI, when this workflow is called with a matrix, GitHub will separate each matrix combination and preserve job ordering within a matrix combination.)