Remove attempt number from artifact
When all jobs are re-run, all v3 and v4 artifacts are deleted.
When some jobs are re-run, v4 artifacts only show on the summary page for the attempt # that they were created. v3 artifacts always show on the latest attempt, even if the job they were created by does not re-run.