patch(_update_bundle.yaml): Alphabetically sort snap revision list (#243)
## Issue

The workflow mistakenly created PRs for updating the snaps list because
of ordering, even though the revisions itself didn't change. This is
because, for python:

```
> {"packages": ["snap1", "snap2", "snap3"]} != {"packages": ["snap1", "snap3", "snap2"]}
True
```
then, `snaps != old_snaps` was being set to true every time, and the
output of `updates_available` was mistakenly being set to `true`
## Solution

Sort the snaps list. See test runs: 

postgresql:
https://github.com/canonical/postgresql-bundle/commit/d5056b5f706d554ad65c45e13be39200a0d82437
mysql:
https://github.com/canonical/mysql-bundle/commit/4269833c7c0e9fcb01407d5c93e93671688de616