breaking: Support multi-charm repos in build & release (#267)
Breaking changes:
- artifact-prefix format changed. Now, if build_*.yaml is called with a
matrix (e.g. using different `path-to-charm-directory` inputs), the
`artifact-prefix` output will be identical. This is to workaround an
issue where if `needs.build.outputs.artifact-prefix` is used on a
`build` job that used a matrix, GitHub picks one of outputs from the
matrix items as the output
- release_charm.yaml and release_snap.yaml now include the charm/snap
name in the git tag for the revision, if the path to charm/snap
directory is not `.`
- `create-github-release` input renamed to `create-git-tags` and now
only git tags are created instead of tags & github release. GitHub
releases should be used for release notes for stable releases and GitHub
releases should be per-repository, not per-charm