breaking(release_python_package.yaml): Only mark GitHub release as latest on default branch (#324)
Breaking change for release_python_package.yaml, _promote_charms.yaml,
and _promote_charm_legacy.yaml

Previously, thought that the latest release was being determined
automatically, per GitHub CLI docs
(https://cli.github.com/manual/gh_release_create)

However, GitHub CLI docs on `--latest` are incorrect; default is to
always mark release as latest

https://github.com/cli/cli/blob/85d6766c621a9d67e0d353a12bb388c40ceca2a1/pkg/cmd/release/create/create.go#L217

https://github.com/cli/cli/blob/85d6766c621a9d67e0d353a12bb388c40ceca2a1/pkg/cmdutil/flags.go#L23
The actual behavior matches the default value for `make_latest` in the
GitHub API:
https://docs.github.com/en/rest/releases/releases?apiVersion=2022-11-28#create-a-release

Instead of relying on the GitHub CLI to determine if the release should
be marked as latest, determine ourselves by checking if the branch that
is being released/promoted is the default branch

Alternatively, we could call the API directly and set `make_latest` to
`legacy`. However, from testing, this appears to only work with semantic
versions (i.e. not charm refresh compatibility versions)—otherwise the
ordering appears to be alphabetical (so it sees "v1.0.0.2" as a newer
version than "v1.0.0.10"). However, this behavior is not documented so
the behavior could be more complex