patch(build_*.yaml): Fix artifact upload
Fixes #329

In release_*.yaml, the craft artifact was downloaded to the incorrect directory if the craft was not located at the root of the repository (e.g. `path-to-rock-directory` was not `.`)

Root cause: actions/upload-artifact 4.4 made a breaking change (without bumping major version, so we didn't see this in their release notes) to exclude hidden files by default: https://github.com/actions/upload-artifact/pull/598

Thank you @gboutry for reporting!